backend/requirements.txt
# Backend dependencies
# requirements.txt

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9  # PostgreSQL adapter (optional)

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Time Series & Forecasting
prophet==1.1.5
statsmodels==0.14.1

# AI/LLM Integration
anthropic>=0.40.0  # Claude API (latest)
openai>=1.50.0     # OpenAI API (optional)

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0